requirements.txt
# Core async
aiohttp==3.9.3
aiofiles==23.2.1

# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.7.0
pydantic-settings==2.3.0

# Database
sqlalchemy==2.0.30
alembic==1.13.1
aiosqlite==0.20.0

# Integrations
python-telegram-bot==21.3
discord.py==2.4.0
slack-bolt==1.18.1

# AI
anthropic==0.28.0

# Git
GitPython==3.1.43
PyGithub==2.3.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
click==8.1.7
loguru==0.7.2
httpx==0.27.0
jinja2==3.1.4

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0

# Code quality
black==24.4.2
ruff==0.4.9
mypy==1.10.0

# Cryptography
cryptography==42.0.8
google-generativeai==0.8.3 .venv/bin/pip install google-generativeai -q
